Add ProductCard component and centralize product data
