+++ data/products.js
@@ -1,0 +1,28 @@
+const products = [
+  {
+    name: 'Island Sunrise',
+    intensity: 6,
+    roast: 'Medium',
+    flavor: 'Bright citrus & vanilla',
+    brand: 'Seychelles Roast',
+    image: '/island-sunrise.jpg'
+  },
+  {
+    name: 'Tropical Midnight',
+    intensity: 9,
+    roast: 'Dark',
+    flavor: 'Cocoa & spice',
+    brand: 'Seychelles Roast',
+    image: '/tropical-midnight.jpg'
+  },
+  {
+    name: 'Coral Espresso',
+    intensity: 8,
+    roast: 'Medium-Dark',
+    flavor: 'Nutty caramel',
+    brand: 'Island Bean',
+    image: '/coral-espresso.jpg'
+  }
+]
+
+export default products
