+++ pages/products.js
@@ -1,28 +1,6 @@
 import Head from 'next/head'
-
-const products = [
-  {
-    name: 'Island Sunrise',
-    intensity: 6,
-    roast: 'Medium',
-    flavor: 'Bright citrus & vanilla',
-    brand: 'Seychelles Roast'
-  },
-  {
-    name: 'Tropical Midnight',
-    intensity: 9,
-    roast: 'Dark',
-    flavor: 'Cocoa & spice',
-    brand: 'Seychelles Roast'
-  },
-  {
-    name: 'Coral Espresso',
-    intensity: 8,
-    roast: 'Medium-Dark',
-    flavor: 'Nutty caramel',
-    brand: 'Island Bean'
-  }
-]
+import ProductCard from '../components/ProductCard'
+import products from '../data/products'
 
 export default function Products() {
   return (
@@ -34,13 +12,7 @@ export default function Products() {
       <h1 className="text-3xl font-semibold mb-6 text-center">Coffee Capsule Selection</h1>
       <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-6">
         {products.map((p) => (
-          <div key={p.name} className="bg-white shadow p-6 rounded">
-            <h2 className="text-xl font-bold mb-2">{p.name}</h2>
-            <p className="text-sm">Intensity: {p.intensity}</p>
-            <p className="text-sm">Roast: {p.roast}</p>
-            <p className="text-sm">Flavor Notes: {p.flavor}</p>
-            <p className="text-sm">Brand: {p.brand}</p>
-          </div>
+          <ProductCard key={p.name} {...p} />
         ))}
       </div>
     </div>
